#4: Controladores CRUD de usuario y asamblea

+++ decisionsystem/src/main/java/es/udc/fic/decisionsystem/repository/sistemaconsulta/SistemaConsultaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import es.udc.fic.decisionsystem.model.sistemaconsulta.SistemaConsulta;
 
 @Repository
-public interface SistemaConsultaRepository extends JpaRepository<SistemaConsulta, Long>{
+public interface SistemaConsultaRepository extends JpaRepository<SistemaConsulta, Integer>{
 
 }
